refactor(kubernautagent): split audit/ds_store.go into focused files
Extracts the per-event-type buildXXXPayload functions into ds_payloads.go,
the data-extraction helpers and IncidentResponseData mapping functions
(toIncidentResponseData, toRemediationTargetResponse, mapSeverity, etc.)
into ds_response_mapping.go, and BufferedDSAuditStore into
ds_buffered_store.go. Reduces ds_store.go from 1035 to 130 lines
(NewDSAuditStore + StoreAudit + buildEventData dispatch only), no
behavioral changes. gofmt fixed incidental alignment/indentation in two
test files matched by the same glob.
