+++ internal/kubernautagent/audit/ds_store.go
@@ -18,13 +18,9 @@ package audit
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"time"
 
-	"github.com/go-faster/jx"
-	"github.com/go-logr/logr"
-	sharedaudit "github.com/jordigilh/kubernaut/pkg/audit"
 	ogenclient "github.com/jordigilh/kubernaut/pkg/datastorage/ogen-client"
 )
 
@@ -132,904 +128,3 @@ func buildEventData(event *AuditEvent) (ogenclient.AuditEventRequestEventData, b
 	}
 	return builder(event), true
 }
-
-func buildEnrichmentCompletedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentEnrichmentCompletedPayload{
-		EventType:                 ogenclient.AIAgentEnrichmentCompletedPayloadEventTypeAiagentEnrichmentCompleted,
-		EventID:                   dataString(event.Data, "event_id"),
-		IncidentID:                dataString(event.Data, "incident_id"),
-		RootOwnerKind:             dataString(event.Data, "root_owner_kind"),
-		RootOwnerName:             dataString(event.Data, "root_owner_name"),
-		OwnerChainLength:          dataInt(event.Data, "owner_chain_length"),
-		RemediationHistoryFetched: dataBool(event.Data, "remediation_history_fetched"),
-	}
-	if ns := dataString(event.Data, "root_owner_namespace"); ns != "" {
-		payload.RootOwnerNamespace.SetTo(ns)
-	}
-	return ogenclient.NewAIAgentEnrichmentCompletedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildEnrichmentFailedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentEnrichmentFailedPayload{
-		EventType:            ogenclient.AIAgentEnrichmentFailedPayloadEventTypeAiagentEnrichmentFailed,
-		EventID:              dataString(event.Data, "event_id"),
-		IncidentID:           dataString(event.Data, "incident_id"),
-		Reason:               dataString(event.Data, "reason"),
-		Detail:               dataString(event.Data, "detail"),
-		AffectedResourceKind: dataString(event.Data, "affected_resource_kind"),
-		AffectedResourceName: dataString(event.Data, "affected_resource_name"),
-	}
-	if ns := dataString(event.Data, "affected_resource_namespace"); ns != "" {
-		payload.AffectedResourceNamespace.SetTo(ns)
-	}
-	return ogenclient.NewAIAgentEnrichmentFailedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildLLMRequestPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.LLMRequestPayload{
-		EventType:     ogenclient.LLMRequestPayloadEventTypeAiagentLlmRequest,
-		EventID:       dataString(event.Data, "event_id"),
-		IncidentID:    event.CorrelationID,
-		Model:         dataString(event.Data, "model"),
-		PromptLength:  dataInt(event.Data, "prompt_length"),
-		PromptPreview: truncate(dataString(event.Data, "prompt_preview"), previewMaxLen),
-	}
-	if tools := dataStringSlice(event.Data, "toolsets_enabled"); len(tools) > 0 {
-		payload.ToolsetsEnabled = tools
-	}
-	return ogenclient.NewLLMRequestPayloadAuditEventRequestEventData(payload)
-}
-
-func buildLLMResponsePayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.LLMResponsePayload{
-		EventType:       ogenclient.LLMResponsePayloadEventTypeAiagentLlmResponse,
-		EventID:         dataString(event.Data, "event_id"),
-		IncidentID:      event.CorrelationID,
-		HasAnalysis:     dataBool(event.Data, "has_analysis"),
-		AnalysisLength:  dataInt(event.Data, "analysis_length"),
-		AnalysisPreview: truncate(dataString(event.Data, "analysis_preview"), previewMaxLen),
-	}
-	if full := dataString(event.Data, "analysis_full"); full != "" {
-		payload.AnalysisFull.SetTo(full)
-	}
-	if tokens := dataInt(event.Data, "total_tokens"); tokens > 0 {
-		payload.TokensUsed.SetTo(tokens)
-	}
-	if tc := dataInt(event.Data, "tool_call_count"); tc > 0 {
-		payload.ToolCallCount.SetTo(tc)
-	}
-	return ogenclient.NewLLMResponsePayloadAuditEventRequestEventData(payload)
-}
-
-func buildLLMToolCallPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.LLMToolCallPayload{
-		EventType:     ogenclient.LLMToolCallPayloadEventTypeAiagentLlmToolCall,
-		EventID:       dataString(event.Data, "event_id"),
-		IncidentID:    event.CorrelationID,
-		ToolCallIndex: dataInt(event.Data, "tool_call_index"),
-		ToolName:      dataString(event.Data, "tool_name"),
-	}
-	if args := dataString(event.Data, "tool_arguments"); args != "" {
-		payload.ToolArguments.SetTo(toJxRawMap(args))
-	}
-	payload.ToolResult = toJxRaw(dataString(event.Data, "tool_result"))
-	if len(payload.ToolResult) == 0 {
-		payload.ToolResult = jx.Raw(`""`)
-	}
-	if preview := dataString(event.Data, "tool_result_preview"); preview != "" {
-		payload.ToolResultPreview.SetTo(truncate(preview, previewMaxLen))
-	}
-	return ogenclient.NewLLMToolCallPayloadAuditEventRequestEventData(payload)
-}
-
-func buildValidationAttemptPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.WorkflowValidationPayload{
-		EventType:   ogenclient.WorkflowValidationPayloadEventTypeAiagentWorkflowValidationAttempt,
-		EventID:     dataString(event.Data, "event_id"),
-		IncidentID:  event.CorrelationID,
-		Attempt:     dataInt(event.Data, "attempt"),
-		MaxAttempts: dataInt(event.Data, "max_attempts"),
-		IsValid:     dataBool(event.Data, "is_valid"),
-	}
-	if errs, ok := event.Data["errors"].([]string); ok {
-		payload.Errors = errs
-	}
-	if wfID := dataString(event.Data, "workflow_id"); wfID != "" {
-		payload.WorkflowID.SetTo(wfID)
-	}
-	if dataBool(event.Data, "is_final_attempt") {
-		payload.IsFinalAttempt.SetTo(true)
-	}
-	return ogenclient.NewWorkflowValidationPayloadAuditEventRequestEventData(payload)
-}
-
-func buildResponseCompletePayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentResponsePayload{
-		EventType:  ogenclient.AIAgentResponsePayloadEventTypeAiagentResponseComplete,
-		EventID:    dataString(event.Data, "event_id"),
-		IncidentID: event.CorrelationID,
-	}
-	if rd := dataString(event.Data, "response_data"); rd != "" {
-		payload.ResponseData = toIncidentResponseData(rd, event.CorrelationID)
-	}
-	if pt := dataInt(event.Data, "total_prompt_tokens"); pt > 0 {
-		payload.TotalPromptTokens.SetTo(pt)
-	}
-	if ct := dataInt(event.Data, "total_completion_tokens"); ct > 0 {
-		payload.TotalCompletionTokens.SetTo(ct)
-	}
-	return ogenclient.NewAIAgentResponsePayloadAuditEventRequestEventData(payload)
-}
-
-func buildRCACompletePayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentRCACompletePayload{
-		EventType:  ogenclient.AIAgentRCACompletePayloadEventTypeAiagentRcaComplete,
-		EventID:    dataString(event.Data, "event_id"),
-		IncidentID: event.CorrelationID,
-	}
-	if rd := dataString(event.Data, "response_data"); rd != "" {
-		payload.ResponseData = toIncidentResponseData(rd, event.CorrelationID)
-	}
-	if pt := dataInt(event.Data, "total_prompt_tokens"); pt > 0 {
-		payload.TotalPromptTokens.SetTo(pt)
-	}
-	if ct := dataInt(event.Data, "total_completion_tokens"); ct > 0 {
-		payload.TotalCompletionTokens.SetTo(ct)
-	}
-	return ogenclient.NewAIAgentRCACompletePayloadAuditEventRequestEventData(payload)
-}
-
-func buildResponseFailedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentResponseFailedPayload{
-		EventType:    ogenclient.AIAgentResponseFailedPayloadEventTypeAiagentResponseFailed,
-		EventID:      dataString(event.Data, "event_id"),
-		IncidentID:   event.CorrelationID,
-		ErrorMessage: dataString(event.Data, "error_message"),
-		Phase:        dataString(event.Data, "phase"),
-	}
-	if dur := dataFloat64(event.Data, "duration_seconds"); dur > 0 {
-		payload.DurationSeconds.SetTo(float32(dur))
-	}
-	return ogenclient.NewAIAgentResponseFailedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionStartedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionStartedPayload{
-		EventType: ogenclient.AIAgentSessionStartedPayloadEventTypeAiagentSessionStarted,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	if v := dataString(event.Data, "incident_id"); v != "" {
-		payload.IncidentID.SetTo(v)
-	}
-	if v := dataString(event.Data, "signal_name"); v != "" {
-		payload.SignalName.SetTo(v)
-	}
-	if v := dataString(event.Data, "severity"); v != "" {
-		payload.Severity.SetTo(v)
-	}
-	if v := dataString(event.Data, "created_by"); v != "" {
-		payload.CreatedBy.SetTo(v)
-	}
-	return ogenclient.NewAIAgentSessionStartedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionCompletedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionCompletedPayload{
-		EventType: ogenclient.AIAgentSessionCompletedPayloadEventTypeAiagentSessionCompleted,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	return ogenclient.NewAIAgentSessionCompletedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionFailedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionFailedPayload{
-		EventType: ogenclient.AIAgentSessionFailedPayloadEventTypeAiagentSessionFailed,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	if v := dataString(event.Data, "error"); v != "" {
-		payload.Error.SetTo(v)
-	}
-	return ogenclient.NewAIAgentSessionFailedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionCancelledPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionCancelledPayload{
-		EventType: ogenclient.AIAgentSessionCancelledPayloadEventTypeAiagentSessionCancelled,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	return ogenclient.NewAIAgentSessionCancelledPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionObservedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionObservedPayload{
-		EventType: ogenclient.AIAgentSessionObservedPayloadEventTypeAiagentSessionObserved,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	if v := dataString(event.Data, "observer_user"); v != "" {
-		payload.ObserverUser.SetTo(v)
-	}
-	if v := dataString(event.Data, "session_owner"); v != "" {
-		payload.SessionOwner.SetTo(v)
-	}
-	return ogenclient.NewAIAgentSessionObservedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionAccessDeniedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionAccessDeniedPayload{
-		EventType:      ogenclient.AIAgentSessionAccessDeniedPayloadEventTypeAiagentSessionAccessDenied,
-		EventID:        dataString(event.Data, "event_id"),
-		SessionID:      event.SessionID,
-		Endpoint:       dataString(event.Data, "endpoint"),
-		RequestingUser: dataString(event.Data, "requesting_user"),
-	}
-	if v := dataString(event.Data, "session_owner"); v != "" {
-		payload.SessionOwner.SetTo(v)
-	}
-	return ogenclient.NewAIAgentSessionAccessDeniedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildInvestigationCancelledPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentInvestigationCancelledPayload{
-		EventType:       ogenclient.AIAgentInvestigationCancelledPayloadEventTypeAiagentInvestigationCancelled,
-		EventID:         dataString(event.Data, "event_id"),
-		CancelledPhase:  dataString(event.Data, "cancelled_phase"),
-		CancelledAtTurn: dataInt(event.Data, "cancelled_at_turn"),
-	}
-	if event.SessionID != "" {
-		payload.SessionID.SetTo(event.SessionID)
-	}
-	if pt := dataInt(event.Data, "total_prompt_tokens"); pt > 0 {
-		payload.TotalPromptTokens.SetTo(pt)
-	}
-	if ct := dataInt(event.Data, "total_completion_tokens"); ct > 0 {
-		payload.TotalCompletionTokens.SetTo(ct)
-	}
-	if tt := dataInt(event.Data, "total_tokens"); tt > 0 {
-		payload.TotalTokens.SetTo(tt)
-	}
-	if v := dataString(event.Data, "accumulated_messages"); v != "" {
-		payload.AccumulatedMessages.SetTo(v)
-	}
-	return ogenclient.NewAIAgentInvestigationCancelledPayloadAuditEventRequestEventData(payload)
-}
-
-func buildAlignmentStepPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentAlignmentStepPayload{
-		EventType:   ogenclient.AIAgentAlignmentStepPayloadEventTypeAiagentAlignmentStep,
-		EventID:     dataString(event.Data, "event_id"),
-		StepIndex:   dataInt(event.Data, "step_index"),
-		StepKind:    dataString(event.Data, "step_kind"),
-		Explanation: dataString(event.Data, "explanation"),
-	}
-	if v := dataString(event.Data, "tool"); v != "" {
-		payload.Tool.SetTo(v)
-	}
-	return ogenclient.NewAIAgentAlignmentStepPayloadAuditEventRequestEventData(payload)
-}
-
-func buildAlignmentVerdictPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentAlignmentVerdictPayload{
-		EventType: ogenclient.AIAgentAlignmentVerdictPayloadEventTypeAiagentAlignmentVerdict,
-		EventID:   dataString(event.Data, "event_id"),
-		Result:    dataString(event.Data, "result"),
-		Flagged:   dataInt(event.Data, "flagged"),
-		Total:     dataInt(event.Data, "total"),
-	}
-	if v := dataString(event.Data, "summary"); v != "" {
-		payload.Summary.SetTo(v)
-	}
-	if pt := dataInt(event.Data, "shadow_prompt_tokens"); pt > 0 {
-		payload.ShadowPromptTokens.SetTo(pt)
-	}
-	if ct := dataInt(event.Data, "shadow_completion_tokens"); ct > 0 {
-		payload.ShadowCompletionTokens.SetTo(ct)
-	}
-	if tt := dataInt(event.Data, "shadow_total_tokens"); tt > 0 {
-		payload.ShadowTotalTokens.SetTo(tt)
-	}
-	return ogenclient.NewAIAgentAlignmentVerdictPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionSuspendedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionSuspendedPayload{
-		EventType: ogenclient.AIAgentSessionSuspendedPayloadEventTypeAiagentSessionSuspended,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	return ogenclient.NewAIAgentSessionSuspendedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSessionResumedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentSessionResumedPayload{
-		EventType: ogenclient.AIAgentSessionResumedPayloadEventTypeAiagentSessionResumed,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	return ogenclient.NewAIAgentSessionResumedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildInteractiveStartedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentInteractiveStartedPayload{
-		EventType: ogenclient.AIAgentInteractiveStartedPayloadEventTypeAiagentInteractiveStarted,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	return ogenclient.NewAIAgentInteractiveStartedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildInteractiveCompletedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentInteractiveCompletedPayload{
-		EventType: ogenclient.AIAgentInteractiveCompletedPayloadEventTypeAiagentInteractiveCompleted,
-		EventID:   dataString(event.Data, "event_id"),
-		SessionID: event.SessionID,
-	}
-	if reason, ok := event.Data["reason"].(string); ok {
-		payload.Reason.SetTo(reason)
-	}
-	return ogenclient.NewAIAgentInteractiveCompletedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildInteractiveK8sCallPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentInteractiveK8sCallPayload{
-		EventType:  ogenclient.AIAgentInteractiveK8sCallPayloadEventTypeAiagentInteractiveK8sCall,
-		EventID:    dataString(event.Data, "event_id"),
-		SessionID:  event.SessionID,
-		ActingUser: event.ActingUser,
-		Resource:   dataString(event.Data, "resource"),
-		Verb:       dataString(event.Data, "verb"),
-	}
-	if ns := dataString(event.Data, "namespace"); ns != "" {
-		payload.Namespace.SetTo(ns)
-	}
-	if name := dataString(event.Data, "resource_name"); name != "" {
-		payload.ResourceName.SetTo(name)
-	}
-	if code := dataInt(event.Data, "http_status_code"); code > 0 {
-		payload.HTTPStatusCode.SetTo(code)
-	}
-	if cid := event.CorrelationID; cid != "" {
-		payload.CorrelationID.SetTo(cid)
-	}
-	return ogenclient.NewAIAgentInteractiveK8sCallPayloadAuditEventRequestEventData(payload)
-}
-
-func buildSecretAccessedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	verb := ogenclient.AIAgentSecretAccessedPayloadVerbGet
-	if dataString(event.Data, "verb") == "list" {
-		verb = ogenclient.AIAgentSecretAccessedPayloadVerbList
-	}
-	payload := ogenclient.AIAgentSecretAccessedPayload{
-		EventType: ogenclient.AIAgentSecretAccessedPayloadEventTypeAiagentSecretAccessed,
-		EventID:   dataString(event.Data, "event_id"),
-		Verb:      verb,
-	}
-	if ns := dataString(event.Data, "namespace"); ns != "" {
-		payload.Namespace.SetTo(ns)
-	}
-	if name := dataString(event.Data, "secret_name"); name != "" {
-		payload.SecretName.SetTo(name)
-	}
-	if toolName := dataString(event.Data, "tool_name"); toolName != "" {
-		payload.ToolName.SetTo(toolName)
-	}
-	if detail := dataString(event.Data, "outcome_detail"); detail != "" {
-		payload.OutcomeDetail.SetTo(detail)
-	}
-	return ogenclient.NewAIAgentSecretAccessedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildShadowLLMRequestPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.ShadowLLMRequestPayload{
-		EventType:    ogenclient.ShadowLLMRequestPayloadEventTypeAiagentShadowLlmRequest,
-		EventID:      dataString(event.Data, "event_id"),
-		IncidentID:   event.CorrelationID,
-		StepIndex:    dataInt(event.Data, "step_index"),
-		StepKind:     dataString(event.Data, "step_kind"),
-		PromptLength: dataInt(event.Data, "prompt_length"),
-	}
-	return ogenclient.NewShadowLLMRequestPayloadAuditEventRequestEventData(payload)
-}
-
-func buildShadowLLMResponsePayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.ShadowLLMResponsePayload{
-		EventType:        ogenclient.ShadowLLMResponsePayloadEventTypeAiagentShadowLlmResponse,
-		EventID:          dataString(event.Data, "event_id"),
-		IncidentID:       event.CorrelationID,
-		StepIndex:        dataInt(event.Data, "step_index"),
-		StepKind:         dataString(event.Data, "step_kind"),
-		PromptTokens:     dataInt(event.Data, "prompt_tokens"),
-		CompletionTokens: dataInt(event.Data, "completion_tokens"),
-		TotalTokens:      dataInt(event.Data, "total_tokens"),
-	}
-	if attempt := dataInt(event.Data, "attempt"); attempt > 0 {
-		payload.Attempt.SetTo(attempt)
-	}
-	if er := dataString(event.Data, "evaluation_result"); er != "" {
-		payload.EvaluationResult.SetTo(ogenclient.ShadowLLMResponsePayloadEvaluationResult(er))
-	}
-	return ogenclient.NewShadowLLMResponsePayloadAuditEventRequestEventData(payload)
-}
-
-func buildRateLimitDeniedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentRatelimitDeniedPayload{
-		EventType: ogenclient.AIAgentRatelimitDeniedPayloadEventTypeAiagentRatelimitDenied,
-		EventID:   dataString(event.Data, "event_id"),
-	}
-	if v := dataString(event.Data, "source_ip"); v != "" {
-		payload.SourceIP.SetTo(v)
-	}
-	if v := dataString(event.Data, "path"); v != "" {
-		payload.Path.SetTo(v)
-	}
-	if v := dataString(event.Data, "method"); v != "" {
-		payload.Method.SetTo(v)
-	}
-	return ogenclient.NewAIAgentRatelimitDeniedPayloadAuditEventRequestEventData(payload)
-}
-
-func buildAuthFailurePayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentAuthFailurePayload{
-		EventType: ogenclient.AIAgentAuthFailurePayloadEventTypeAiagentAuthFailure,
-		EventID:   dataString(event.Data, "event_id"),
-	}
-	if v := dataString(event.Data, "source_ip"); v != "" {
-		payload.SourceIP.SetTo(v)
-	}
-	if v := dataString(event.Data, "path"); v != "" {
-		payload.Path.SetTo(v)
-	}
-	if v := dataString(event.Data, "method"); v != "" {
-		payload.Method.SetTo(v)
-	}
-	return ogenclient.NewAIAgentAuthFailurePayloadAuditEventRequestEventData(payload)
-}
-
-func buildAuthDeniedPayload(event *AuditEvent) ogenclient.AuditEventRequestEventData {
-	payload := ogenclient.AIAgentAuthDeniedPayload{
-		EventType: ogenclient.AIAgentAuthDeniedPayloadEventTypeAiagentAuthDenied,
-		EventID:   dataString(event.Data, "event_id"),
-	}
-	if v := dataString(event.Data, "source_ip"); v != "" {
-		payload.SourceIP.SetTo(v)
-	}
-	if v := dataString(event.Data, "path"); v != "" {
-		payload.Path.SetTo(v)
-	}
-	if v := dataString(event.Data, "method"); v != "" {
-		payload.Method.SetTo(v)
-	}
-	return ogenclient.NewAIAgentAuthDeniedPayloadAuditEventRequestEventData(payload)
-}
-
-func dataString(d map[string]interface{}, key string) string {
-	if v, ok := d[key]; ok {
-		if s, ok := v.(string); ok {
-			return s
-		}
-	}
-	return ""
-}
-
-func dataInt(d map[string]interface{}, key string) int {
-	if v, ok := d[key]; ok {
-		switch n := v.(type) {
-		case int:
-			return n
-		case int64:
-			return int(n)
-		case float64:
-			return int(n)
-		}
-	}
-	return 0
-}
-
-func dataBool(d map[string]interface{}, key string) bool {
-	if v, ok := d[key]; ok {
-		if b, ok := v.(bool); ok {
-			return b
-		}
-	}
-	return false
-}
-
-func dataFloat64(d map[string]interface{}, key string) float64 {
-	if v, ok := d[key]; ok {
-		switch n := v.(type) {
-		case float64:
-			return n
-		case float32:
-			return float64(n)
-		case int:
-			return float64(n)
-		}
-	}
-	return 0
-}
-
-func dataStringSlice(d map[string]interface{}, key string) []string {
-	if v, ok := d[key]; ok {
-		if ss, ok := v.([]string); ok {
-			return ss
-		}
-	}
-	return nil
-}
-
-const previewMaxLen = 500
-
-func truncate(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	runes := []rune(s)
-	if len(runes) <= maxLen {
-		return s
-	}
-	return string(runes[:maxLen])
-}
-
-func toJxRaw(s string) jx.Raw {
-	if s != "null" && json.Valid([]byte(s)) {
-		return jx.Raw(s)
-	}
-	b, err := json.Marshal(s)
-	if err != nil {
-		return jx.Raw(`""`)
-	}
-	return jx.Raw(b)
-}
-
-func toJxRawMap(s string) ogenclient.LLMToolCallPayloadToolArguments {
-	var raw map[string]json.RawMessage
-	if err := json.Unmarshal([]byte(s), &raw); err != nil {
-		return nil
-	}
-	result := make(ogenclient.LLMToolCallPayloadToolArguments, len(raw))
-	for k, v := range raw {
-		result[k] = jx.Raw(v)
-	}
-	return result
-}
-
-type investigationResultJSON struct {
-	RCASummary           string                 `json:"rca_summary"`
-	Severity             string                 `json:"severity"`
-	ContributingFactors  []string               `json:"contributing_factors"`
-	WorkflowID           string                 `json:"workflow_id"`
-	ExecutionBundle      string                 `json:"execution_bundle"`
-	Confidence           float64                `json:"confidence"`
-	NeedsHumanReview     bool                   `json:"needs_human_review"`
-	HumanReviewReason    string                 `json:"human_review_reason"`
-	Warnings             []string               `json:"warnings"`
-	Parameters           map[string]interface{} `json:"parameters"`
-	AlternativeWorkflows []altWorkflowJSON      `json:"alternative_workflows"`
-	RemediationTarget    *remediationTargetJSON `json:"remediation_target"`
-	CausalChain          []string               `json:"causal_chain"`
-	DueDiligence         *dueDiligenceJSON      `json:"due_diligence"`
-}
-
-type dueDiligenceJSON struct {
-	CausalCompleteness    string `json:"causal_completeness"`
-	TargetAccuracy        string `json:"target_accuracy"`
-	EvidenceSufficiency   string `json:"evidence_sufficiency"`
-	AlternativeHypotheses string `json:"alternative_hypotheses"`
-	ScopeCompleteness     string `json:"scope_completeness"`
-	Proportionality       string `json:"proportionality"`
-	RegressionAwareness   string `json:"regression_awareness"`
-	ConfidenceCalibration string `json:"confidence_calibration"`
-}
-
-type altWorkflowJSON struct {
-	WorkflowID      string  `json:"workflow_id"`
-	Rationale       string  `json:"rationale"`
-	ExecutionBundle string  `json:"execution_bundle"`
-	Confidence      float64 `json:"confidence"`
-}
-
-type remediationTargetJSON struct {
-	Kind      string `json:"kind"`
-	Name      string `json:"name"`
-	Namespace string `json:"namespace"`
-}
-
-func toIncidentResponseData(responseDataJSON string, incidentID string) ogenclient.IncidentResponseData {
-	var ir investigationResultJSON
-	if err := json.Unmarshal([]byte(responseDataJSON), &ir); err != nil {
-		return ogenclient.IncidentResponseData{
-			IncidentId: incidentID,
-			Timestamp:  time.Now().UTC(),
-		}
-	}
-
-	data := ogenclient.IncidentResponseData{
-		IncidentId: incidentID,
-		Analysis:   ir.RCASummary,
-		Confidence: float32(ir.Confidence),
-		Timestamp:  time.Now().UTC(),
-		RootCauseAnalysis: ogenclient.IncidentResponseDataRootCauseAnalysis{
-			Summary:             ir.RCASummary,
-			Severity:            mapSeverity(ir.Severity),
-			ContributingFactors: ir.ContributingFactors,
-		},
-	}
-
-	if ir.RemediationTarget != nil {
-		data.RootCauseAnalysis.RemediationTarget.SetTo(toRemediationTargetResponse(ir.RemediationTarget))
-	}
-
-	if len(ir.CausalChain) > 0 {
-		data.RootCauseAnalysis.CausalChain = ir.CausalChain
-	}
-	if ir.DueDiligence != nil {
-		data.RootCauseAnalysis.DueDiligence.SetTo(toDueDiligenceResponse(ir.DueDiligence))
-	}
-
-	if ir.WorkflowID != "" {
-		data.SelectedWorkflow.SetTo(toSelectedWorkflowResponse(ir))
-	}
-
-	if ir.NeedsHumanReview {
-		data.NeedsHumanReview.SetTo(true)
-	}
-	if ir.HumanReviewReason != "" {
-		if reason, ok := validHumanReviewReason(ir.HumanReviewReason); ok {
-			data.HumanReviewReason.SetTo(reason)
-		}
-	}
-	if len(ir.Warnings) > 0 {
-		data.Warnings = ir.Warnings
-	} else {
-		data.Warnings = []string{}
-	}
-
-	if len(ir.AlternativeWorkflows) > 0 {
-		data.AlternativeWorkflows = toAlternativeWorkflowsResponse(ir.AlternativeWorkflows)
-	}
-
-	return data
-}
-
-// toRemediationTargetResponse maps the audit-record RemediationTarget into
-// the ogen wire format, omitting any field that was never populated.
-func toRemediationTargetResponse(rt *remediationTargetJSON) ogenclient.IncidentResponseDataRootCauseAnalysisRemediationTarget {
-	var out ogenclient.IncidentResponseDataRootCauseAnalysisRemediationTarget
-	if rt.Kind != "" {
-		out.Kind.SetTo(rt.Kind)
-	}
-	if rt.Name != "" {
-		out.Name.SetTo(rt.Name)
-	}
-	if rt.Namespace != "" {
-		out.Namespace.SetTo(rt.Namespace)
-	}
-	return out
-}
-
-// toDueDiligenceResponse maps the audit-record DueDiligence review into the
-// ogen wire format, omitting any field that was never populated.
-func toDueDiligenceResponse(dd *dueDiligenceJSON) ogenclient.IncidentResponseDataRootCauseAnalysisDueDiligence {
-	out := ogenclient.IncidentResponseDataRootCauseAnalysisDueDiligence{}
-	if dd.CausalCompleteness != "" {
-		out.CausalCompleteness.SetTo(dd.CausalCompleteness)
-	}
-	if dd.TargetAccuracy != "" {
-		out.TargetAccuracy.SetTo(dd.TargetAccuracy)
-	}
-	if dd.EvidenceSufficiency != "" {
-		out.EvidenceSufficiency.SetTo(dd.EvidenceSufficiency)
-	}
-	if dd.AlternativeHypotheses != "" {
-		out.AlternativeHypotheses.SetTo(dd.AlternativeHypotheses)
-	}
-	if dd.ScopeCompleteness != "" {
-		out.ScopeCompleteness.SetTo(dd.ScopeCompleteness)
-	}
-	if dd.Proportionality != "" {
-		out.Proportionality.SetTo(dd.Proportionality)
-	}
-	if dd.RegressionAwareness != "" {
-		out.RegressionAwareness.SetTo(dd.RegressionAwareness)
-	}
-	if dd.ConfidenceCalibration != "" {
-		out.ConfidenceCalibration.SetTo(dd.ConfidenceCalibration)
-	}
-	return out
-}
-
-// toSelectedWorkflowResponse maps the selected-workflow fields of ir into the
-// ogen wire format. Callers must check ir.WorkflowID != "".
-func toSelectedWorkflowResponse(ir investigationResultJSON) ogenclient.IncidentResponseDataSelectedWorkflow {
-	sw := ogenclient.IncidentResponseDataSelectedWorkflow{}
-	sw.WorkflowId.SetTo(ir.WorkflowID)
-	if ir.ExecutionBundle != "" {
-		sw.ExecutionBundle.SetTo(ir.ExecutionBundle)
-	}
-	sw.Confidence.SetTo(float32(ir.Confidence))
-	if len(ir.Parameters) > 0 {
-		params := make(ogenclient.IncidentResponseDataSelectedWorkflowParameters, len(ir.Parameters))
-		for k, v := range ir.Parameters {
-			b, err := json.Marshal(v)
-			if err != nil {
-				continue
-			}
-			params[k] = jx.Raw(b)
-		}
-		sw.Parameters.SetTo(params)
-	}
-	return sw
-}
-
-// toAlternativeWorkflowsResponse maps the audit-record AlternativeWorkflows
-// list into the ogen wire format.
-func toAlternativeWorkflowsResponse(alts []altWorkflowJSON) []ogenclient.IncidentResponseDataAlternativeWorkflowsItem {
-	out := make([]ogenclient.IncidentResponseDataAlternativeWorkflowsItem, 0, len(alts))
-	for _, alt := range alts {
-		item := ogenclient.IncidentResponseDataAlternativeWorkflowsItem{}
-		if alt.WorkflowID != "" {
-			item.WorkflowId.SetTo(alt.WorkflowID)
-		}
-		if alt.Rationale != "" {
-			item.Rationale.SetTo(alt.Rationale)
-		}
-		if alt.ExecutionBundle != "" {
-			item.ExecutionBundle.SetTo(alt.ExecutionBundle)
-		}
-		if alt.Confidence > 0 {
-			item.Confidence.SetTo(float32(alt.Confidence))
-		}
-		out = append(out, item)
-	}
-	return out
-}
-
-func mapSeverity(s string) ogenclient.IncidentResponseDataRootCauseAnalysisSeverity {
-	switch s {
-	case "critical":
-		return ogenclient.IncidentResponseDataRootCauseAnalysisSeverityCritical
-	case "high":
-		return ogenclient.IncidentResponseDataRootCauseAnalysisSeverityHigh
-	case "warning":
-		return ogenclient.IncidentResponseDataRootCauseAnalysisSeverityWarning
-	case "info":
-		return ogenclient.IncidentResponseDataRootCauseAnalysisSeverityInfo
-	default:
-		return ogenclient.IncidentResponseDataRootCauseAnalysisSeverityUnknown
-	}
-}
-
-var validHumanReviewReasons = map[string]ogenclient.IncidentResponseDataHumanReviewReason{
-	"workflow_not_found":          ogenclient.IncidentResponseDataHumanReviewReasonWorkflowNotFound,
-	"image_mismatch":              ogenclient.IncidentResponseDataHumanReviewReasonImageMismatch,
-	"parameter_validation_failed": ogenclient.IncidentResponseDataHumanReviewReasonParameterValidationFailed,
-	"no_matching_workflows":       ogenclient.IncidentResponseDataHumanReviewReasonNoMatchingWorkflows,
-	"low_confidence":              ogenclient.IncidentResponseDataHumanReviewReasonLowConfidence,
-	"llm_parsing_error":           ogenclient.IncidentResponseDataHumanReviewReasonLlmParsingError,
-	"investigation_inconclusive":  ogenclient.IncidentResponseDataHumanReviewReasonInvestigationInconclusive,
-	"rca_incomplete":              ogenclient.IncidentResponseDataHumanReviewReasonRcaIncomplete,
-	"alignment_check_failed":      ogenclient.IncidentResponseDataHumanReviewReasonAlignmentCheckFailed,
-}
-
-// validHumanReviewReason returns the ogen enum value if the string is a recognised
-// HumanReviewReason. Free-text values from the LLM are rejected so that the
-// ogen client-side OpenAPI validation does not drop the entire audit event.
-func validHumanReviewReason(s string) (ogenclient.IncidentResponseDataHumanReviewReason, bool) {
-	v, ok := validHumanReviewReasons[s]
-	return v, ok
-}
-
-var _ AuditStore = (*DSAuditStore)(nil)
-
-// BufferedDSAuditStore wraps pkg/audit.BufferedAuditStore to implement KA's
-// internal AuditStore interface. Events are converted from KA's AuditEvent
-// format to the OpenAPI AuditEventRequest format, then enqueued into the
-// platform buffered store for batched writes with retry.
-//
-// Uses the same OpenAPIClientAdapter + BufferedAuditStore stack as every other
-// platform service (DD-AUDIT-002 alignment).
-type BufferedDSAuditStore struct {
-	inner sharedaudit.AuditStore
-}
-
-// BufferedDSAuditStoreOption allows callers to override RecommendedConfig fields.
-type BufferedDSAuditStoreOption func(*sharedaudit.Config)
-
-// WithFlushInterval overrides the default flush interval from RecommendedConfig.
-func WithFlushInterval(d time.Duration) BufferedDSAuditStoreOption {
-	return func(c *sharedaudit.Config) {
-		if d > 0 {
-			c.FlushInterval = d
-		}
-	}
-}
-
-// WithBufferSize overrides the default buffer size from RecommendedConfig.
-func WithBufferSize(n int) BufferedDSAuditStoreOption {
-	return func(c *sharedaudit.Config) {
-		if n > 0 {
-			c.BufferSize = n
-		}
-	}
-}
-
-// WithBatchSize overrides the default batch size from RecommendedConfig.
-func WithBatchSize(n int) BufferedDSAuditStoreOption {
-	return func(c *sharedaudit.Config) {
-		if n > 0 {
-			c.BatchSize = n
-		}
-	}
-}
-
-// NewBufferedDSAuditStore creates a KA audit store backed by the platform
-// BufferedAuditStore. The caller provides a DataStorageClient (typically
-// created via audit.NewOpenAPIClientAdapterWithTransport to share auth
-// transport with the rest of KA's DS operations).
-//
-// Optional BufferedDSAuditStoreOption values override the RecommendedConfig
-// defaults, allowing integration/E2E tests to control flush behaviour via
-// the same YAML fields that KA v1.2 used (flush_interval_seconds, etc.).
-func NewBufferedDSAuditStore(dsClient sharedaudit.DataStorageClient, logger logr.Logger, opts ...BufferedDSAuditStoreOption) (*BufferedDSAuditStore, error) {
-	cfg := sharedaudit.RecommendedConfig("kubernaut-agent")
-	for _, o := range opts {
-		o(&cfg)
-	}
-	inner, err := sharedaudit.NewBufferedStore(dsClient, cfg, "kubernaut-agent", logger)
-	if err != nil {
-		return nil, fmt.Errorf("create buffered audit store: %w", err)
-	}
-	return &BufferedDSAuditStore{inner: inner}, nil
-}
-
-func (s *BufferedDSAuditStore) StoreAudit(ctx context.Context, event *AuditEvent) error {
-	req := &ogenclient.AuditEventRequest{
-		Version:        "1.0",
-		EventType:      event.EventType,
-		EventTimestamp: time.Now().UTC(),
-		EventCategory:  ogenclient.AuditEventRequestEventCategory(event.EventCategory),
-		EventAction:    event.EventAction,
-		EventOutcome:   ogenclient.AuditEventRequestEventOutcome(event.EventOutcome),
-		CorrelationID:  event.CorrelationID,
-	}
-	if event.ActingUser != "" {
-		req.ActorType.SetTo("User")
-		req.ActorID.SetTo(event.ActingUser)
-	} else {
-		bActorType := "Service"
-		bActorID := "kubernaut-agent"
-		if event.ActorID != "" {
-			bActorID = event.ActorID
-		}
-		if event.ActorType != "" {
-			bActorType = event.ActorType
-		}
-		req.ActorType.SetTo(bActorType)
-		req.ActorID.SetTo(bActorID)
-	}
-	if event.ClusterName != "" {
-		req.ClusterName.SetTo(event.ClusterName)
-	}
-
-	if ed, ok := buildEventData(event); ok {
-		req.EventData = ed
-	}
-
-	return s.inner.StoreAudit(ctx, req)
-}
-
-// Flush forces all buffered events to be written to DataStorage.
-func (s *BufferedDSAuditStore) Flush(ctx context.Context) error {
-	return s.inner.Flush(ctx)
-}
-
-// Close flushes remaining events and stops the background worker.
-func (s *BufferedDSAuditStore) Close() error {
-	return s.inner.Close()
-}
-
-var _ AuditStore = (*BufferedDSAuditStore)(nil)
