+++ internal/kubernautagent/audit/ds_store_test.go
@@ -204,15 +204,15 @@ var _ = Describe("Kubernaut Agent DS Audit Store — TP-433-WIR Phase 7", func()
 				event.Data["flagged"] = 0
 				event.Data["total"] = 5
 
-			// Interactive events do not have dedicated OpenAPI discriminator variants;
-			// they are stored with outer fields only (event_type, session_id, etc.).
-			recorder.calls = nil
-			err := store.StoreAudit(context.Background(), event)
-			Expect(err).NotTo(HaveOccurred(), "StoreAudit should succeed for %s", eventType)
-			Expect(recorder.calls).To(HaveLen(1), "should record call for %s", eventType)
-			Expect(recorder.calls[0].EventData.Type).NotTo(BeZero(),
-				"buildEventData should return a non-zero type for %s", eventType)
-		}
+				// Interactive events do not have dedicated OpenAPI discriminator variants;
+				// they are stored with outer fields only (event_type, session_id, etc.).
+				recorder.calls = nil
+				err := store.StoreAudit(context.Background(), event)
+				Expect(err).NotTo(HaveOccurred(), "StoreAudit should succeed for %s", eventType)
+				Expect(recorder.calls).To(HaveLen(1), "should record call for %s", eventType)
+				Expect(recorder.calls[0].EventData.Type).NotTo(BeZero(),
+					"buildEventData should return a non-zero type for %s", eventType)
+			}
 		})
 	})
